chore: projectsNav tailwind migration
