+++ package.json
@@ -18,6 +18,7 @@
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "classnames": "^2.5.1",
     "commitlint": "^17.6.5",
     "eslint-plugin-tailwindcss": "^3.13.1",
     "i18next": "^23.2.0",
