+++ src/components/bars/index.ts
@@ -1,3 +1,3 @@
 export * from "./navbar";
-export * from "./projects-nav/projects-nav";
+export * from "./projects-nav";
 export * from "./projects-infos/projects-infos";
