--- src/components/bars/projects-nav/styles.module.scss
@@ -1,76 +1,0 @@
-@import "/src/assets/styles/variables.scss";
-
-.projectsNav {
-  width: 150px;
-  height: calc(100vh - 150px * 2);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  box-sizing: border-box;
-  border-right-width: 2px;
-  border-right-style: solid;
-  border-image: linear-gradient(
-      180deg,
-      rgba(255, 255, 255, 0) 1.56%,
-      rgba(255, 255, 255, 0.9) 49.48%,
-      rgba(255, 255, 255, 0) 100%
-    )
-    1;
-
-  .content {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 16px;
-
-    .arrow {
-      width: 18px;
-      cursor: pointer;
-      height: auto;
-      user-select: none;
-      opacity: 1;
-      transition: transform 0.7s ease, opacity 0.7s ease;
-      &:hover {
-        opacity: 0.6;
-        transform: scale(0.9);
-        transition: transform 0.7s ease, opacity 0.7s ease;
-      }
-    }
-
-    .arrowDisabled {
-      pointer-events: none;
-      opacity: 0.1;
-      transition: opacity 0.7s ease;
-    }
-
-    .projects {
-      display: flex;
-      flex-direction: column;
-      align-items: center;
-      gap: 8px;
-      width: fit-content;
-
-      .divider {
-        width: 40%;
-        height: 1px;
-        background-color: $color-p-black;
-      }
-
-      .current {
-        font-family: "Cinzel Decorative";
-        font-size: 26px;
-        font-weight: bold;
-        color: $color-p-black;
-        line-height: 1.5;
-      }
-
-      .total {
-        font-family: "Cinzel Decorative";
-        font-size: 26px;
-        font-weight: 400;
-        color: $color-p-black;
-        line-height: 1.5;
-      }
-    }
-  }
-}
